Fixed Issue with Turnip
HasLineOfSight, when used with an object that is interpolated, will cause issues on the validity of the raycast. However, this can be fixed by changing the length of the raycast to the Vector3 distance and instead raycasting forever, we just raycast the distance see if we hit either the player or nothing. As long as a wall was not hit, then we should be fine.